Move apps out of site
