+++ src/aas/alert/apps.py
@@ -2,4 +2,5 @@ from django.apps import AppConfig
 
 
 class AlertConfig(AppConfig):
-    name = "aas.site.alert"
+    name = "aas.alert"
+    label = 'aas_alert'
