+++ src/aas/auth/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = "aas.auth.apps.AuthConfig"
