+++ src/aas/notificationprofile/apps.py
@@ -2,4 +2,5 @@ from django.apps import AppConfig
 
 
 class NotificationprofileConfig(AppConfig):
-    name = "aas.site.notificationprofile"
+    name = "aas.notificationprofile"
+    label = "aas_notoprofile"
