+++ src/aas/notificationprofile/notification_media.py
@@ -7,9 +7,9 @@ from django.conf import settings
 from django.template.loader import render_to_string
 from rest_framework.renderers import JSONRenderer
 
-from aas.site.alert.models import Alert
-from aas.site.alert.serializers import AlertSerializer
-from aas.site.auth.models import User
+from aas.alert.models import Alert
+from aas.alert.serializers import AlertSerializer
+from aas.auth.models import User
 from .models import NotificationProfile
