+++ src/aas/notificationprofile/templates/notificationprofile/email.txt
