--- src/aas/site/auth/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = "aas.site.auth.apps.AuthConfig"
